requirements.txt
flask
langchain
langchain-community
pinecone-client
requests
aiohttp
asyncio
python-dotenv
PyMuPDF
pandas
Pillow
tabula-py
base64
gunicorn